requirements.txt
# External dependencies for the project

# Optional (for colored output)
coloramagit 